+++ packages/@expo/cli/src/utils/env.ts
@@ -214,6 +214,11 @@ class Env {
   get __EXPO_EAGER_BUNDLE_OPTIONS() {
     return string('__EXPO_EAGER_BUNDLE_OPTIONS', '');
   }
+
+  /** Disable server deployment during production builds (during `expo export:embed`). This is useful for testing API routes and server components against a local server. */
+  get EXPO_NO_DEPLOY(): boolean {
+    return boolish('EXPO_NO_DEPLOY', false);
+  }
 }
 
 export const env = new Env();